+++ products/TreeComponentFramework.browser.js
@@ -1269,6 +1269,9 @@ class AbstractTreeComponentRootNode extends AbstractTreeComponent {
   async appWillFirstRender() {}
   // todo: remove?
   async appDidFirstRender() {}
+  onCommandError(err) {
+    throw new err()
+  }
   _setTreeComponentFrameworkEventListeners() {
     const willowBrowser = this.getWillowProgram()
     const bodyShadow = willowBrowser.getBodyStumpNode().getShadow()
@@ -1280,7 +1283,11 @@ class AbstractTreeComponentRootNode extends AbstractTreeComponent {
       const commandWithArgs = stumpNode.getStumpNodeAttr(attr)
       const commandArgs = commandWithArgs.split(" ")
       const command = commandArgs.shift()
-      commander[command](...commandArgs)
+      try {
+        commander[command](...commandArgs)
+      } catch (err) {
+        this.onCommandError(err)
+      }
       return false
     }
     const DataShadowEvents = WillowConstants.DataShadowEvents